+++ splink/internals/em_training_session.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
 import logging
+from dataclasses import asdict, dataclass
 from typing import TYPE_CHECKING, List
 
 from splink.internals.blocking import BlockingRule, block_using_rules_sqls
 from splink.internals.charts import (
-    ChartReturnType,
-    m_u_parameters_interactive_history_chart,
-    match_weights_interactive_history_chart,
-    probability_two_random_records_match_iteration_chart,
+    MatchWeightsInteractiveHistoryChart,
+    MUParametersInteractiveHistoryChart,
+    ProbabilityTwoRandomRecordsMatchIterationChart,
 )
 from splink.internals.comparison import Comparison
 from splink.internals.comparison_vector_values import (
@@ -22,6 +22,7 @@ from splink.internals.pipeline import CTEPipeline
 from splink.internals.settings import (
     ComparisonAndLevelDict,
     CoreModelSettings,
+    ModelParameterDetailedRecord,
     Settings,
     TrainingSettings,
 )
@@ -37,6 +38,27 @@ logger = logging.getLogger(__name__)
 if TYPE_CHECKING:
     from splink.internals.linker import Linker
     from splink.internals.splink_dataframe import SplinkDataFrame
+
+
+@dataclass
+class ModelParameterIterationDetailedRecord(ModelParameterDetailedRecord):
+    iteration: int
+
+    @classmethod
+    def from_settings_param_detailed_record(
+        cls,
+        cl_rec: ModelParameterDetailedRecord,
+        *,
+        probability_two_random_records_match: float,
+        iteration: int,
+    ) -> ModelParameterIterationDetailedRecord:
+        cl_rec.probability_two_random_records_match = (
+            probability_two_random_records_match
+        )
+        return cls(
+            **asdict(cl_rec),
+            iteration=iteration,
+        )
 
 
 class EMTrainingSession:
@@ -320,20 +342,20 @@ class EMTrainingSession:
         return adjusted_prop_m
 
     @property
-    def _iteration_history_records(self):
+    def _iteration_history_records(self) -> list[ModelParameterIterationDetailedRecord]:
         output_records = []
 
         for iteration, core_model_settings in enumerate(
             self._core_model_settings_history
         ):
-            records = core_model_settings.parameters_as_detailed_records
-
-            for r in records:
-                r["iteration"] = iteration
-                # TODO: why lambda from current settings, not history?
-                r["probability_two_random_records_match"] = (
-                    self.core_model_settings.probability_two_random_records_match
+            records = [
+                ModelParameterIterationDetailedRecord.from_settings_param_detailed_record(
+                    r,
+                    probability_two_random_records_match=self.core_model_settings.probability_two_random_records_match,
+                    iteration=iteration,
                 )
+                for r in core_model_settings.parameters_as_detailed_records
+            ]
 
             output_records.extend(records)
         return output_records
@@ -352,7 +374,9 @@ class EMTrainingSession:
             output_records.append(r)
         return output_records
 
-    def probability_two_random_records_match_iteration_chart(self) -> ChartReturnType:
+    def probability_two_random_records_match_iteration_chart(
+        self,
+    ) -> ProbabilityTwoRandomRecordsMatchIterationChart:
         """
         Display a chart showing the iteration history of the probability that two
         random records match.
@@ -361,29 +385,32 @@ class EMTrainingSession:
             An interactive Altair chart.
         """
         records = self._lambda_history_records
-        return probability_two_random_records_match_iteration_chart(records)
+        return ProbabilityTwoRandomRecordsMatchIterationChart(records)
 
-    def match_weights_interactive_history_chart(self) -> ChartReturnType:
+    def match_weights_interactive_history_chart(
+        self,
+    ) -> MatchWeightsInteractiveHistoryChart:
         """
         Display an interactive chart of the match weights history.
 
         Returns:
             An interactive Altair chart.
         """
-        records = self._iteration_history_records
-        return match_weights_interactive_history_chart(
-            records, blocking_rule=self._blocking_rule_for_training.blocking_rule_sql
+        return MatchWeightsInteractiveHistoryChart(
+            self._iteration_history_records,
+            blocking_rule_text=self._blocking_rule_for_training.blocking_rule_sql,
         )
 
-    def m_u_values_interactive_history_chart(self) -> ChartReturnType:
+    def m_u_values_interactive_history_chart(
+        self,
+    ) -> MUParametersInteractiveHistoryChart:
         """
         Display an interactive chart of the m and u values.
 
         Returns:
             An interactive Altair chart.
         """
-        records = self._iteration_history_records
-        return m_u_parameters_interactive_history_chart(records)
+        return MUParametersInteractiveHistoryChart(self._iteration_history_records)
 
     def __repr__(self):
         deactivated_cols = ", ".join(