+++ splink/internals/spark/dataframe.py
@@ -44,6 +44,14 @@ class SparkDataFrame(SplinkDataFrame):
 
         return [r.asDict(recursive=True) for r in spark_df.collect()]
 
+    def as_dict(self, limit=None):
+        sql = f"select * from {self.physical_name}"
+        if limit:
+            sql += f" limit {limit}"
+
+        spark_df = self.db_api._execute_sql_against_backend(sql)
+        return {col: spark_df.select(col).rdd.collect() for col in spark_df.columns}
+
     def as_pyarrow_table(self, limit: int = None) -> PyArrowTable:
         # spark 3 doesn't have native arrow support, so use our fallback method instead
         if get_spark_major_version() == 3: